+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Release notes
 - Improve the stability of the "Check for new Package versions" feature.
   https://github.com/nexB/dejacode/issues/17
 
+- Improve the support for SourgeForge download URLs.
+  https://github.com/nexB/dejacode/issues/26
+
 ### Version 5.0.0
 
 Initial release.